Moved new API's to util
